Add Unlicense to project
This commit adds the Unlicense to the project, allowing anyone to copy, modify, publish, use, compile, sell, or distribute the software for any purpose. The authors dedicate all copyright interest in the software to the public domain. The software is provided "as is" without warranty. For more information, please refer to http://unlicense.org